fix(karrotframe): remove all singletons
