--- packages/karrotframe/src/navigator/helpers/generateScreenInstanceId.ts
@@ -1,6 +1,0 @@
-let screenInstanceIdCount = 0
-
-export function generateScreenInstanceId() {
-  screenInstanceIdCount += 1
-  return String(screenInstanceIdCount)
-}
