+++ packages/karrotframe/src/navigator/helpers/index.ts
@@ -1,3 +1,2 @@
-export * from './generateScreenInstanceId'
 export * from './getNavigatorParams'
 export * from './NavigatorTheme'
